feat!: run flottbot as unprivileged user in docker (#142)
* Run flottbot as unprivileged user in docker

Runs flottbot as an unprivileged user rather than root in the docker
images. Adds flottbot as a system user with uid 900 by default but
allows setting those values as env vars and set the image to use that
user at runtime.

* Update Dockerfiles to explictly set flottbot group

Updates the Dockerfiles to explictly set up a flottbot group and
add the flottbot user to that as adduser in Alpine doesn't do that
by default.